[camel-k] move SyndesisHeaderStrategy to integration/runtime #5189
