+++ app/integration/runtime-camelk/pom.xml
@@ -62,10 +62,6 @@
       <groupId>org.apache.camel</groupId>
       <artifactId>camel-core</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.apache.camel</groupId>
-      <artifactId>camel-http-common</artifactId>
-    </dependency>
 
     <!-- set atlasmap as always required runtime dependency-->
     <dependency>
