+++ app/integration/runtime/src/test/java/io/syndesis/integration/runtime/util/SyndesisHeaderStrategyTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.syndesis.connector.support.processor;
+package io.syndesis.integration.runtime.util;
 
 import org.apache.camel.Exchange;
 import org.junit.Test;
